+++ plugins/modal-resize.js
@@ -187,17 +187,19 @@
 	};
 
 	var getHorizontalOffset = function () {
+		var innerWidth = global.innerWidth || document.body.clientWidth;
 		var element = CSSModal.activeElement.querySelector('.modal-inner');
 		var elementWidth = parseInt(global.getComputedStyle(element).getPropertyValue('width'), 10);
-		var offset = (global.innerWidth - elementWidth) / 2;
+		var offset = (innerWidth - elementWidth) / 2;
 
 		return offset;
 	};
 
 	var getVerticalOffset = function () {
+		var innerHeight = global.innerHeight || document.body.clientHeight;
 		var element = CSSModal.activeElement.querySelector('.modal-inner');
 		var elementHeight = parseInt(global.getComputedStyle(element).getPropertyValue('height'), 10);
-		var offset = (global.innerHeight - elementHeight) / 2;
+		var offset = (innerHeight - elementHeight) / 2;
 
 		return offset;
 	};