+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add gallery thumbnail images to gallery footer
 * Always show gallery arrows on mobile
 
+* Add compatiblity for IE8 (use "bean" library for event handling)
 * Allow the content with id 0 to be selected
 
 * Fix max height of modal content